fix: recommendations from vitvakatu
